pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gangge-code"
version = "0.1.0"
description = "AI Coding Assistant — 5-Layer Architecture"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "textual>=0.80.0",
    "rich>=13.9.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.10.0",
    "watchdog>=5.0.0",
    "pyperclip>=1.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.7.0",
]

[project.scripts]
gangge = "gangge.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
